+++ agent/src/test/java/com/teamscale/jacoco/agent/options/AgentOptionsTest.java
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.io.TempDir;
 import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.nio.file.Path;
 import java.util.function.Predicate;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -188,60 +187,6 @@ public class AgentOptionsTest {
 		AgentOptions agentOptions = getAgentOptionsParserWithDummyLogger()
 				.parse("excludes=" + filterString);
 		return string -> agentOptions.getLocationIncludeFilter().isIncluded(string);
-	}
-
-	/** Tests path resolution with absolute path. */
-	@Test
-	public void testPathResolutionForAbsolutePath() throws AgentOptionParseException {
-		assertInputInWorkingDirectoryMatches(".", testFolder.getAbsolutePath(), "");
-	}
-
-	/** Tests path resolution with relative paths. */
-	@Test
-	public void testPathResolutionForRelativePath() throws AgentOptionParseException {
-		assertInputInWorkingDirectoryMatches(".", ".", "");
-		assertInputInWorkingDirectoryMatches("plugins", "../file_with_manifest1.jar", "file_with_manifest1.jar");
-	}
-
-	/** Tests path resolution with patterns and relative paths. */
-	@Test
-	public void testPathResolutionWithPatternsAndRelativePaths() throws AgentOptionParseException {
-		assertInputInWorkingDirectoryMatches(".", "plugins/file_*.jar", "plugins/file_with_manifest2.jar");
-		assertInputInWorkingDirectoryMatches(".", "*/file_*.jar", "plugins/file_with_manifest2.jar");
-		assertInputInWorkingDirectoryMatches("plugins/inner", "..", "plugins");
-		assertInputInWorkingDirectoryMatches("plugins/inner", "../s*", "plugins/some_other_file.jar");
-	}
-
-	/** Tests path resolution with patterns and absolute paths. */
-	@Test
-	public void testPathResolutionWithPatternsAndAbsolutePaths() throws AgentOptionParseException {
-		assertInputInWorkingDirectoryMatches("plugins", testFolder.getAbsolutePath() + "/plugins/file_*.jar",
-				"plugins/file_with_manifest2.jar");
-	}
-
-	/** Tests path resolution with incorrect input. */
-	@Test
-	public void testPathResolutionWithPatternErrorCases() {
-		final File workingDirectory = testFolder;
-		assertThatThrownBy(
-				() -> getAgentOptionsParserWithDummyLogger().parsePath("option-name", workingDirectory, "**.war"))
-				.isInstanceOf(AgentOptionParseException.class).hasMessageContaining(
-				"Invalid path given for option option-name: " +
-						"**.war. The pattern **.war did not match any files in");
-	}
-
-	private void assertInputInWorkingDirectoryMatches(String workingDir, String input,
-													  String expected) throws AgentOptionParseException {
-		final File workingDirectory = new File(testFolder, workingDir);
-		File actualFile = getAgentOptionsParserWithDummyLogger().parsePath("option-name", workingDirectory, input)
-				.toFile();
-		File expectedFile = new File(testFolder, expected);
-		assertThat(getNormalizedPath(actualFile)).isEqualByComparingTo(getNormalizedPath(expectedFile));
-	}
-
-	/** Resolves the path to its absolute normalized path. */
-	private static Path getNormalizedPath(File file) {
-		return file.getAbsoluteFile().toPath().normalize();
 	}
 
 	private static AgentOptionsParser getAgentOptionsParserWithDummyLogger() {